second commit

+++ Coins_club_owner/views.py
@@ -1,4 +1,6 @@
 
+
+
 from django.http import HttpResponse,JsonResponse
 from Mufo.Minxins import *
 from .serializers import *
@@ -16,17 +18,21 @@ from Jockey_club_owner.models import Jockey_club_owner
 import secrets
 from django.utils.decorators import method_decorator
 from Mufo.Minxins import authenticate_token
-
+from master.serializers import *
+import uuid
 def coins_club_owner(request):
     return HttpResponse("Hello, world. You're at the Coins_club_owner index.")
 
 
 class Register(APIView):
     serializer_class = UserSerializer
+    serializer_class1 = masterSerializer
     parser_classes = (MultiPartParser, FormParser, JSONParser)
 
     def post(self, request):
         serializer = self.serializer_class(data=request.data)
+        serializer1 = self.serializer_class1(data=request.data)
+
         phone=serializer.initial_data.get('phone')
         email=serializer.initial_data.get('email')
         if serializer.is_valid():
@@ -61,7 +67,11 @@ class Register(APIView):
                 return Response({'message': message}, status=status.HTTP_400_BAD_REQUEST)
 
             token = secrets.token_hex(128)
-            serializer.save(token =token)
+            uid = uuid.uuid1()
+            usertype="Coins_club_owner"
+            serializer.save(token =token,uid=uid,usertype=usertype)
+            if serializer1.is_valid():
+                serializer1.save(token =token,uid=uid,usertype=usertype)
             user = Coins_club_owner.objects.get(email=serializer.data['email'])
             messages.add_message(request, messages.INFO, f"New coins club owner {user} is registered. please Approve ")
             return Response({'message': "Register successfully. Please wait for some time to Get Approved."}, status=status.HTTP_201_CREATED)
@@ -87,31 +97,48 @@ class CointraderList(APIView):
 
 
 
+
 class UpdateUser(APIView):
     @method_decorator(authenticate_token)
     def get(self, request, format=None):
-        pk = request.user.id
-        user = Coins_club_owner.objects.get(id=pk)
+        pk = request.user.uid
+        user = Coins_club_owner.objects.get(uid=pk)
         serializer = UserUpdateSerializer(user)
         return Response(serializer.data)
+
     @method_decorator(authenticate_token)
-    def put(self, request,format=None):
-        pk = request.user.id
-        user = Coins_club_owner.objects.get(id=pk)
+    def put(self, request, format=None):
+        uid = request.user.uid
+        print(uid)
+        user = Coins_club_owner.objects.get(uid=uid)
+        print(user)
+        common_objects = Common.objects.get(uid=uid)
+        print(common_objects)
         serializer = UserUpdateSerializer(user, data=request.data)
+        serializer1 = masterUpdateSerializer(common_objects, data=request.data)
+        if common_objects:
+            serializer1 = masterUpdateSerializer(common_objects, data=request.data)
+            if serializer1.is_valid():
+                serializer1.save()                
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
-        else:
-            return Response(serializer.errors)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    
     @method_decorator(authenticate_token)
     def delete(self, request, format=None):
-        pk = request.user.id
-        user = Coins_club_owner.objects.get(id=pk)
-        user.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-
+        pk = request.user.uid
+        print(pk)
+        user = Coins_club_owner.objects.get(uid=pk)
+        print(user)
+        commonuser = Common.objects.get(uid=pk)
+        print(commonuser)
+        if commonuser:
+            commonuser.delete()
+            user.delete()
+            return Response({"delete":"successfully"})
+        return Response({"delete":"unsuccessfully"})
 
 @method_decorator(authenticate_token, name='dispatch')
 class userview(APIView):
@@ -120,3 +147,16 @@ class userview(APIView):
         user = request.user
         print(user)
         return JsonResponse({'uid': user.uid, 'number': user.phone,"name":user.Name})
+    
+
+class Alluser(APIView):
+    def get(self, request):
+        data = Coins_club_owner.objects.all()
+        approved_users = []
+        for user in data:
+            if user.Is_Approved:
+                approved_users.append(user)
+        if approved_users:
+            serializer = UserSerializer(approved_users, many=True)
+            return Response(serializer.data)
+        return Response("No approved users found")
