+++ Coins_trader/views.py
@@ -1,4 +1,6 @@
 
+
+
 from django.http import HttpResponse,JsonResponse
 from Mufo.Minxins import *
 from .serializers import *
@@ -17,7 +19,8 @@ import secrets
 
 from django.utils.decorators import method_decorator
 from Mufo.Minxins import authenticate_token
-
+from master.serializers import *
+import uuid
 
 
 def coins_trader(request):
@@ -26,10 +29,12 @@ def coins_trader(request):
 
 class Register(APIView):
     serializer_class = UserSerializer
+    serializer_class1 = masterSerializer
     parser_classes = (MultiPartParser, FormParser, JSONParser)
 
     def post(self, request):
         serializer = self.serializer_class(data=request.data)
+        serializer1 = self.serializer_class1(data=request.data)
         phone=serializer.initial_data.get('phone')
         email=serializer.initial_data.get('email')
         if serializer.is_valid():
@@ -63,7 +68,11 @@ class Register(APIView):
                 message = 'Email already exists as an Jockey_club_owner ' if email_exists else 'Phone number already exists as an Jockey_club_owner '
                 return Response({'message': message}, status=status.HTTP_400_BAD_REQUEST)
             token = secrets.token_hex(128)
-            serializer.save(token =token)
+            uid = uuid.uuid1()
+            usertype="Coins_trader"
+            serializer.save(token =token,uid=uid,usertype=usertype)
+            if serializer1.is_valid():
+                serializer1.save(token =token,uid=uid,usertype=usertype)
             user = Coins_trader.objects.get(email=serializer.data['email'])
             messages.add_message(request, messages.INFO, f"New Coins Trader {user} is registered. please Approve ")
             return Response({'message': "Register successfully. Please wait for some time to Get Approved."}, status=status.HTTP_201_CREATED)
@@ -91,33 +100,42 @@ class CointraderConnectedOwner(APIView):
         except Coins_trader.DoesNotExist:
             return Response({'message': 'coin trader not found'}, status=status.HTTP_404_NOT_FOUND)
 
-
-
 class UpdateUser(APIView):
     @method_decorator(authenticate_token)
     def get(self, request, format=None):
-        pk = request.user.id
-        user = Coins_trader.objects.get(id=pk)
+        pk = request.user.uid
+        user = Coins_trader.objects.get(uid=pk)
         serializer = UserUpdateSerializer(user)
         return Response(serializer.data)
+
     @method_decorator(authenticate_token)
-    def put(self, request,format=None):
-        pk = request.user.id
-        user = Coins_trader.objects.get(id=pk)
+    def put(self, request, format=None):
+        uid = request.user.uid
+        user = Coins_trader.objects.get(uid=uid)
+        common_objects = Common.objects.get(uid=uid)
         serializer = UserUpdateSerializer(user, data=request.data)
+        serializer1 = masterUpdateSerializer(common_objects, data=request.data)
+        if common_objects:
+            serializer1 = masterUpdateSerializer(common_objects, data=request.data)
+            if serializer1.is_valid():
+                serializer1.save()                
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
-        else:
-            return Response(serializer.errors)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    
     @method_decorator(authenticate_token)
     def delete(self, request, format=None):
-        pk = request.user.id
-        user = Coins_trader.objects.get(id=pk)
-        user.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-
+        pk = request.user.uid
+        user = Coins_trader.objects.get(uid=pk)
+        commonuser = Common.objects.get(uid=pk)
+        if commonuser:
+            commonuser.delete()
+            user.delete()
+            return Response({"delete":"successfully"})
+        return Response({"delete":"unsuccessfully"})
+ 
 
 @method_decorator(authenticate_token, name='dispatch')
 class userview(APIView):
@@ -126,3 +144,16 @@ class userview(APIView):
         user = request.user
         print(user)
         return JsonResponse({'uid': user.uid, 'number': user.phone,"name":user.Name})
+
+class Alluser(APIView):
+    def get(self, request):
+        data = Coins_trader.objects.all()
+        
+        approved_users = []
+        for user in data:
+            if user.Is_Approved:
+                approved_users.append(user)
+        if approved_users:
+            serializer = UserSerializer(approved_users, many=True)
+            return Response(serializer.data)
+        return Response("No approved users found")
